color: split out ANSI parsing code
Move the pure ANSI parsing functions into a separate file
so that they can be tested.

- ansi_color_parse_single()
- ansi_color_seq_length()
- ansi_is_end_char()
- ansi_skip_sequence()
