+++ Makefile.autosetup
@@ -181,7 +181,8 @@ $(PWD)/browser:
 # libcolor
 LIBCOLOR=	libcolor.a
 LIBCOLOROBJS=	color/ansi.o color/attr.o color/color.o color/command.o \
-		color/curses.o color/merged.o color/notify.o color/quoted.o \
+		color/curses.o color/merged.o color/notify.o \
+		color/parse_ansi.o color/quoted.o \
 		color/regex.o color/simple.o
 @if USE_DEBUG_COLOR
 LIBCOLOROBJS+=	color/debug.o
