+++ color/lib.h
@@ -35,6 +35,7 @@
  * | color/debug.c       | @subpage color_debug       |
  * | color/merged.c      | @subpage color_merge       |
  * | color/notify.c      | @subpage color_notify      |
+ * | color/parse_ansi.c  | @subpage color_parse_ansi  |
  * | color/quoted.c      | @subpage color_quote       |
  * | color/regex.c       | @subpage color_regex       |
  * | color/simple.c      | @subpage color_simple      |
@@ -52,6 +53,7 @@
 #include "debug.h"
 #include "merged.h"
 #include "notify2.h"
+#include "parse_ansi.h"
 #include "quoted.h"
 #include "regex4.h"
 #include "simple2.h"
