+++ color/parse_ansi.h
@@ -1,0 +1,35 @@
+/**
+ * @file
+ * Parse ANSI Sequences
+ *
+ * @authors
+ * Copyright (C) 2023 Richard Russon <rich@flatcap.org>
+ *
+ * @copyright
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 2 of the License, or (at your option) any later
+ * version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef MUTT_COLOR_PARSE_ANSI_H
+#define MUTT_COLOR_PARSE_ANSI_H
+
+#include <stdbool.h>
+
+struct AnsiColor;
+
+int ansi_color_parse_single(const char *buf, struct AnsiColor *ansi, bool dry_run);
+int ansi_color_seq_length  (const char *str);
+
+#endif /* MUTT_COLOR_PARSE_ANSI_H */
+
+
