+++ debug/lib.h
@@ -43,6 +43,7 @@
 
 #include <stddef.h>
 #include <stdbool.h>
+#include <time.h>
 #include "mutt/lib.h"
 #include "email/lib.h"
 #include "core/lib.h"
